Initializes an empty Ember.js application
Why is this change needed?
--------------------------
We want to do some experimentation. We are going to build out a simple
table component. The table component will display a list of files that
can be selected and subsequently downloaded.

We will be utilizing the Ember.js JavaScript framework.

How does it address the issue?
------------------------------
This uses the ember-cli package to initialize an empty Ember.js app
using the command below, which creates all the files necessary to get
the app up and running:
```
ember init --yarn --no-welcome
```

We've chosen to use Yarn as our package manager, and we're going to go
ahead and opt-out of creating a welcome component.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://cli.emberjs.com/release/
https://yarnpkg.com