Fix Node.js version to 14.15.4
Why is this change needed?
--------------------------
It's possible to receive the following warning when using a version of
Node.js that the Ember CLI has not be tested against:
```
WARNING: Node v15.7.0 is not tested against Ember CLI on your platform.
```

Additionally, in order to prevent inconsistencies across environments we
should use a consistent node version.

How does it address the issue?
------------------------------
The EmberCLI recommends using "the most-recent 'Active LTS' version of
Node.js". This creates a .node-version file in the root directory that
tools like nodenv can utilize to select a consistent node version.

We've selected 14.15.4 as it is the most recent version of Node.js the
Ember CLI has been officially tested against.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/nodenv/nodenv
https://github.com/ember-cli/ember-cli/blob/master/docs/node-support.md