Implement a static mockup of the table UI
Why is this change needed?
--------------------------
We want to start writing tests and identify potential landmarks for
componentization.

How does it address the issue?
------------------------------
We've first added a few acceptance tests to validate the expected
functionality users will encounter at the root URL of our app. We've
added a single test initially to ensure some version of our table
displays. Several additional tests have been stubbed out with
descriptions of the functionality we intend to build.

Additionally, this builds out a **rough** version of the table UI. The
semantics of the HTML are still a draft and the architecture of the CSS
is sure to change as things become componentized.

To move things along quickly, we've dumped this into the application
template.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/simplabs/qunit-dom/blob/HEAD/API.md