Create <DataTable /> component and refactor styles
Why is this change needed?
--------------------------
Having all of this static html in the application template was a good
starting point, but refactoring it into reusable components would be
better.

How does it address the issue?
------------------------------
This creates a static data-table component template. This follows the
pod-based Ember.js architecture. The biggest benefits achieved here are:
  1. All the corresponding files will be co-located, making the files
     quicker to locate and easier to update or even remove _(gasp!)_.
  2. As the app grows a less flat architecture generally keeps things a
     little more tidy. Iteration becomes more efficient.

With the creation of this component the corresponding styles have also
been refactored. We've chosen to adopt a loose variation of the BEM
class convention. BEM makes it easier infer how markup is structured and
components are abstracted.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://cli.emberjs.com/release/advanced-use/project-layouts/#podslayout
https://css-tricks.com/bem-101/