Update table component to utilize real data
Why is this change needed?
--------------------------
Up until now, everything has rendered static HTML. The table component
will not be reusable until it is updated to render a dynamic set of
data.

How does it address the issue?
------------------------------
This first abstracts the data into the index route file and passes it to
the data-table component as an argument.

A data-table component class has been added to more reliably parse the
data structure passed in. When no data is passed in the table renders
without content. When an array of objects is passed in, the data-table
component class will first parse through the objects and build an array
of headers before associating those headers with the table rows and
columns.